requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
astroid==1.6.3
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
boto==2.48.0
bs4==0.0.1
celery==4.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.4
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-framework==0.1.0
django-ses==0.8.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
emoji==0.5.0
Emojipedia==0.4.0
future==0.16.0
gunicorn==19.8.1
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.8.4
python-decouple==3.1
pytz==2018.4
PyYAML==3.12
raven==6.7.0
requests==2.18.4
requests-toolbelt==0.8.0
simplegeneric==0.8.1
six==1.11.0
tqdm==4.23.1
traitlets==4.3.2
ua-parser==0.8.0
urllib3==1.22
user-agents==1.1.0
vine==1.1.4
wcwidth==0.1.7
wrapt==1.10.11
